refactor(security): extract CSP generation into internal/security/csp
Six csp_* files plus security_config_csp_audit.go move to internal/security/csp.

That last file is the one place filename grouping and the reference graph
disagree. Its name files it with the config cluster; nothing in the config
cluster has ever referenced it. CSPAudit is part of the CSP response schema —
its only reader is CSPResponse.Audit — so it moves with the schema it belongs
to, as csp/audit.go. policy is now three files and knows nothing about CSP.

De-stuttered inside the new package: CSPGenerator -> Generator, NewCSPGenerator
-> NewGenerator, CSPParams/CSPResponse/CSPObservations/CSPAudit -> Params/
Response/Observations/Audit, GenerateCSP -> Generate. HandleGenerateCSP keeps
its name: it is the method capture.CSPGenerator declares and it mirrors the
generate_csp tool name. No JSON tag changes.

The session-override half of security_boundary_test.go and the CSP branch tests
stranded in security_coverage_part2_test.go move here. csp is at 98.0% coverage
on its own tests, httpsec at 100%.
