+++ internal/security/csp/tooling.go
@@ -1,6 +1,7 @@
+// tooling.go — Session whitelist overrides and the MCP generate_csp adapter.
 // Purpose: Applies session-scoped whitelist overrides and integrates CSP generation with MCP tools.
 // Why: Separates tool integration and override application from core CSP generation.
-package security
+package csp
 
 import (
 	"encoding/json"
@@ -15,7 +16,7 @@ import (
 //
 // Failure semantics:
 // - Overrides affect generated output only; no persistent config is mutated.
-func (g *CSPGenerator) applyWhitelistOverrides(response *CSPResponse, overrides []string) {
+func (g *Generator) applyWhitelistOverrides(response *Response, overrides []string) {
 	if response.Directives["default-src"] == nil {
 		response.Directives["default-src"] = []string{"'self'"}
 	}
@@ -40,7 +41,7 @@ func (g *CSPGenerator) applyWhitelistOverrides(response *CSPResponse, overrides 
 		))
 	}
 
-	response.Audit = &CSPAudit{
+	response.Audit = &Audit{
 		SessionOverrides:    overrides,
 		PersistentWhitelist: []string{},
 		OverrideSource:      "mcp_tool_parameter",
@@ -59,7 +60,7 @@ func (g *CSPGenerator) applyWhitelistOverrides(response *CSPResponse, overrides 
 
 // RecordOriginFromBody extracts origin and resource type from a NetworkBody
 // and records it in the origin accumulator. Called from the network ingestion path.
-func (g *CSPGenerator) RecordOriginFromBody(body capture.NetworkBody, pageURL string) {
+func (g *Generator) RecordOriginFromBody(body capture.NetworkBody, pageURL string) {
 	origin := util.ExtractOrigin(body.URL)
 	if origin == "" {
 		return
@@ -69,14 +70,14 @@ func (g *CSPGenerator) RecordOriginFromBody(body capture.NetworkBody, pageURL st
 }
 
 // HandleGenerateCSP is the MCP tool handler for generate_csp.
-func (g *CSPGenerator) HandleGenerateCSP(params json.RawMessage) (any, error) {
-	var cspParams CSPParams
+func (g *Generator) HandleGenerateCSP(params json.RawMessage) (any, error) {
+	var cspParams Params
 	if len(params) > 0 {
 		if err := json.Unmarshal(params, &cspParams); err != nil {
 			return nil, fmt.Errorf("invalid CSP parameters: %w", err)
 		}
 	}
 
-	resp := g.GenerateCSP(cspParams)
+	resp := g.Generate(cspParams)
 	return &resp, nil
 }
