+++ internal/security/security_coverage_part2_test.go
@@ -2,10 +2,10 @@
 // Docs: docs/features/feature/security-hardening/index.md
 
 // security_coverage_part2_test.go — Targeted coverage tests for uncovered security paths (part 2).
-// Covers: checkPII integration, checkSecurityHeaders, shouldSkipHSTS, CSP eviction/GetPages,
-// diffSingleCookieFlag, flagAbsentValue, security config, looksLikeCreditCard,
+// Covers: checkPII integration, checkSecurityHeaders, shouldSkipHSTS,
+// diffSingleCookieFlag, flagAbsentValue, looksLikeCreditCard,
 // scanConsoleForCredentials, getEntryString, scanURLForGenericSecrets, SRI,
-// directiveForResourceType, checkTransport, checkCookies, parseSingleCookie.
+// checkTransport, checkCookies.
 package security
 
 import (
@@ -90,54 +90,6 @@ func TestCheckSecurityHeaders_SkipsHSTSForLocalhost(t *testing.T) {
 	}
 	if shouldSkipHSTS("X-Frame-Options", capture.NetworkBody{URL: "http://localhost:3000"}) {
 		t.Error("should not skip non-HSTS header")
-	}
-}
-
-// ============================================
-// CSP Generator — evictOldestOrigin / GetPages
-// ============================================
-
-func TestCSPGenerator_EvictOldestOrigin(t *testing.T) {
-	t.Parallel()
-	g := NewCSPGenerator()
-
-	// Record 10001+ unique origins to trigger eviction
-	for i := 0; i < 10002; i++ {
-		origin := "https://evict-origin-" + strings.Repeat("x", 5) + "-" + padInt(i)
-		g.RecordOrigin(origin, "script", "https://evict-page.example.com")
-	}
-
-	g.mu.RLock()
-	count := len(g.origins)
-	g.mu.RUnlock()
-
-	if count > 10001 {
-		t.Errorf("origin count = %d, should be capped after eviction", count)
-	}
-}
-
-func padInt(n int) string {
-	s := ""
-	for n > 0 {
-		s = string(rune('0'+n%10)) + s
-		n /= 10
-	}
-	if s == "" {
-		return "0"
-	}
-	return s
-}
-
-func TestCSPGenerator_GetPages(t *testing.T) {
-	t.Parallel()
-	g := NewCSPGenerator()
-
-	g.RecordOrigin("https://cdn.example.com", "script", "https://page1.example.com")
-	g.RecordOrigin("https://cdn.example.com", "style", "https://page2.example.com")
-
-	pages := g.GetPages()
-	if len(pages) != 2 {
-		t.Errorf("GetPages() len = %d, want 2", len(pages))
 	}
 }
 
@@ -442,35 +394,6 @@ func TestGenerateTagTemplate_Unknown(t *testing.T) {
 	got := generateTagTemplate("https://cdn.example.com/data.json", "sha384-abc", "data")
 	if got != "" {
 		t.Errorf("expected empty for unknown type, got %q", got)
-	}
-}
-
-// ============================================
-// directiveForResourceType — unknown type fallback
-// ============================================
-
-func TestDirectiveForResourceType_UnknownType(t *testing.T) {
-	t.Parallel()
-	got := directiveForResourceType("unknown")
-	if got != "default-src" {
-		t.Errorf("directiveForResourceType(unknown) = %q, want default-src", got)
-	}
-}
-
-func TestDirectiveForResourceType_KnownTypes(t *testing.T) {
-	t.Parallel()
-	cases := map[string]string{
-		"script":  "script-src",
-		"style":   "style-src",
-		"font":    "font-src",
-		"img":     "img-src",
-		"connect": "connect-src",
-	}
-	for resType, want := range cases {
-		got := directiveForResourceType(resType)
-		if got != want {
-			t.Errorf("directiveForResourceType(%q) = %q, want %q", resType, got, want)
-		}
 	}
 }
